# ticketapp-admin-service/src/main/java/com/ticketapp/dto/PassengerDto.java
package com.ticketapp.dto;

import com.ticketapp.model.enums.Gender;
import lombok.Data;

@Data
public class PassengerDto {
    private Long id;
    private String name;
    private String surname;
    private Gender gender;
}